Change vehicle version when updating

+++ client/src/app/vehicle-catalog/vehicle-catalog.component.ts
@@ -9,6 +9,7 @@ import * as _moment from 'moment';
 import {MatDialog} from "@angular/material/dialog";
 import {DialogVehicleDetailsComponent} from "./dialog-vehicle-details/dialog-vehicle-details.component";
 import {LogInService} from "../api/login-in.service";
+import {MatSnackBar} from "@angular/material/snack-bar";
 
 const moment = _moment;
 
@@ -67,7 +68,9 @@ export class VehicleCatalogComponent implements OnInit {
   constructor(
     private loginService: LogInService,
     private vehicleApiService: VehicleApiService,
-    public dialog: MatDialog) {
+    public dialog: MatDialog,
+    private snackBar: MatSnackBar,
+  ) {
   }
 
   ngOnInit() {
@@ -195,13 +198,20 @@ export class VehicleCatalogComponent implements OnInit {
   }
 
   deleteVehicle(vehicle: Vehicle) {
-    this.isLoading = true;
-    this.vehicleApiService.deleteVehicle(vehicle).subscribe(() => {
-      let index: number = this.dataSource.data.findIndex(d => d === vehicle);
-      this.dataSource.data.splice(index, 1);
-      this.dataSource = new MatTableDataSource<Vehicle>(this.dataSource.data);
-      this.isLoading = false;
-    });
+    this.vehicleApiService.isResourceAvailable(vehicle).subscribe(result => {
+      if (result) {
+        this.isLoading = true;
+        this.vehicleApiService.deleteVehicle(vehicle).subscribe(() => {
+          let index: number = this.dataSource.data.findIndex(d => d === vehicle);
+          this.dataSource.data.splice(index, 1);
+          this.dataSource = new MatTableDataSource<Vehicle>(this.dataSource.data);
+          this.isLoading = false;
+        });
+      } else {
+        this.snackBar.open('Resource unavailable. Try again later', '', {duration: 5000});
+      }
+    })
+
   }
